just for profile

+++ oneflow/user/kernels/pooling_kernel.cpp
@@ -300,8 +300,8 @@ class MaxPool2dKernel final : public user_op::OpKernel {
     const auto* pooling_cache = dynamic_cast<const PoolingOpKernelCache*>(cache);
     const MaxPoolingParams3D& params_3d = pooling_cache->GetParams3D();
 
-    // const int64_t elem_num = y->shape().elem_cnt();
-    const int32_t elem_num = y->shape().elem_cnt();
+    const int64_t elem_num = y->shape().elem_cnt();
+    // const int32_t elem_num = y->shape().elem_cnt();
 
     const T* src = x->dptr<T>();
     T* dest = y->mut_dptr<T>();
